requirements.txt
openai>=1.0.0
httpx>=0.27.0
PyYAML>=6.0.1
python-dotenv>=1.0.1
yfinance>=0.2.54
akshare>=1.16.0
jinja2>=3.1.4
tenacity>=9.0.0
markdown>=3.7
